fix: exporting the client from db itself
