+++ packages/db/src/prisma.ts
@@ -1,0 +1,18 @@
+import { PrismaClient } from '@prisma/client';
+
+const prismaClientSingleton = () => {
+  return new PrismaClient();
+};
+
+type PrismaClientSingleton = ReturnType<typeof prismaClientSingleton>;
+
+// eslint-disable-next-line
+const globalForPrisma = globalThis as unknown as {
+  prisma: PrismaClientSingleton | undefined;
+};
+
+const prisma = globalForPrisma.prisma ?? prismaClientSingleton();
+
+export const db =  prisma;
+
+if (process.env.NODE_ENV !== 'production') globalForPrisma.prisma = prisma;
